Remove SVG toggle option, fixes https://github.com/Codeinwp/otter-blocks/issues/911
