+++ src/blocks/blocks/lottie/editor.scss
@@ -18,9 +18,3 @@
 		margin-top: 0;
 	}
 }
-
-.wp-block-themeisle-blocks-lottie-modal {
-	.components-modal__frame {
-		width: 360px;
-	}
-}
