+++ src/blocks/blocks/lottie/placeholder.js
@@ -15,25 +15,18 @@ import {
 
 import {
 	Button,
-	ExternalLink,
-	Modal,
 	Placeholder
 } from '@wordpress/components';
 
-import {
-	Fragment,
-	useState
-} from '@wordpress/element';
+import { useState } from '@wordpress/element';
 
 const BlockPlaceholder = ({
 	className,
 	file,
 	onChange,
-	isJSONAllowed,
 	attributes
 }) => {
 	const [ url, setURL ] = useState( attributes.file?.url || null );
-	const [ isOpen, setOpen ] = useState( false );
 
 	const onChangeValue = e => {
 		if ( e ) {
@@ -42,25 +35,8 @@ const BlockPlaceholder = ({
 		return onChange( url );
 	};
 
-	if ( isJSONAllowed && ! Boolean( window.themeisleGutenberg.isWPVIP ) ) {
+	if ( Boolean( window.themeisleGutenberg.isWPVIP ) ) {
 		return (
-			<MediaPlaceholder
-				labels={ {
-					title: __( 'Lottie', 'otter-blocks' ),
-					instructions: __( 'Add Lottie animations and files to your website.', 'otter-blocks' )
-				} }
-				icon={ <BlockIcon icon={ video } /> }
-				accept={ [ 'application/json' ] }
-				allowedTypes={ [ 'application/json' ] }
-				value={ { ...file } }
-				onSelectURL={ onChange }
-				onSelect={ onChange }
-			/>
-		);
-	}
-
-	return (
-		<Fragment>
 			<Placeholder
 				label={ __( 'Lottie', 'otter-blocks' ) }
 				instructions={ __( 'Add Lottie animations and files to your website.', 'otter-blocks' ) }
@@ -70,7 +46,7 @@ const BlockPlaceholder = ({
 				<form onSubmit={ onChangeValue }>
 					<input
 						type="url"
-						value={ url }
+						value={ url || '' }
 						className="components-placeholder__input"
 						aria-label={ __( 'Lottie', 'otter-blocks' ) }
 						placeholder={ __( 'Enter URL to embed here…', 'otter-blocks' ) }
@@ -84,33 +60,24 @@ const BlockPlaceholder = ({
 					>
 						{ __( 'Embed', 'otter-blocks' ) }
 					</Button>
-
-					{ ! Boolean( window.themeisleGutenberg.isWPVIP ) && (
-						<Button
-							isSecondary
-							onClick={ () => setOpen( true ) }
-						>
-							{ __( 'Upload', 'otter-blocks' ) }
-						</Button>
-					) }
 				</form>
 			</Placeholder>
+		);
+	}
 
-			{ isOpen && (
-				<Modal
-					title={ __( 'Allow JSON Uploads', 'otter-blocks' ) }
-					closeLabel={ __( 'Close' ) }
-					onRequestClose={ () => setOpen( false ) }
-					overlayClassName="wp-block-themeisle-blocks-lottie-modal"
-				>
-					{ __( 'This file type is not permitted for security reasons. Would you still like to enable JSON uploads?', 'otter-blocks' ) }
-					<br /><br />
-					<ExternalLink href={ window.themeisleGutenberg.optionsPath }>{ __( 'You can enable JSON uploads from Otter.', 'otter-blocks' ) }</ExternalLink>
-					<br /><br />
-					{ __( 'You will have to refresh the page after changing JSON upload settings.', 'otter-blocks' ) }
-				</Modal>
-			) }
-		</Fragment>
+	return (
+		<MediaPlaceholder
+			labels={ {
+				title: __( 'Lottie', 'otter-blocks' ),
+				instructions: __( 'Add Lottie animations and files to your website.', 'otter-blocks' )
+			} }
+			icon={ <BlockIcon icon={ video } /> }
+			accept={ [ 'application/json' ] }
+			allowedTypes={ [ 'application/json' ] }
+			value={ { ...file } }
+			onSelectURL={ onChange }
+			onSelect={ onChange }
+		/>
 	);
 };
